requirements.txt
PyGithub==2.3.0
sentence-transformers==3.0.1
anthropic==0.34.2
streamlit==1.38.0
plotly==5.24.1
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.5.2
python-dotenv==1.0.1
